COMMUNICATION-370 - DLMS: correctly name the serverAddress
The serverUpper- and serverLowerMacAddress are now obsolete. Both properties are combined in 1 -> ServerMacAddress. The structure of this property depends on the connectionMode.
If HDLC is used then the property can contain an upper- and a lowerAddress separated by a colon.
If IP is used then only the upperAddress is taken into consideration.



git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@45717 0d108963-8a47-0410-809b-ebd9fb3b3407