COMMUNICATION-292 - ACE4000 protocol
Added implementation for property "MeterType".
1 = DC meter
2 = CT-meter.

This has an impact on the unit scale for certain registers.

Fixed the broken DT tag!!
Timesync now happens correctly.
It remains however IMPOSSIBLE to check the time difference and base a sync / force time on this.

Fixed the timeout property, max = 40000, otherwise the meter doesn't seem to reply anymore

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@50857 0d108963-8a47-0410-809b-ebd9fb3b3407