IskraMx372 Small bugfix to CSDCall.
When the CSDCall schedule is executed (just making a call), then the protocol is executed. This should be just a dummy - doing nothing.
In this case, when the MeterSerialNumber is requested, the SerialNumber of the RTU (from protocolProperties) will be returned, instead of empty string. --> verifySerialNumber() method of SmartMeterProtocolTaskExecutor will not complain about invalid / mismatch in serial numbers.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@59779 0d108963-8a47-0410-809b-ebd9fb3b3407