EISERVERSG-1195 - Convert Actaris ACE4000 Generic Protocol to a new DeviceProtocol
Ported the old Generic ACE4000 protocol to the new DeviceProtocol format.
It is split into 3 parts: the MBus protocol (simply a place holder, has no extra functionality), the inbound protocol (does a discover and reads in extra data) and the outbound protocol.

The protocol is no longer asynchronous: after a specific request, we read in frames until we receive the expected response (or a timeout occurs).
Some requests (LP, events) have a response that can consists of multiple frames, in this case we read in frames until a timeout occurs and then we parse all received frames. We do this because it is unknown how much frames the meter will send.
Added a protocol dialect for ACE4000 (consists of only timeout and retry property)
Added NoSecuritySupport implementation

The same protocol base can be used for SMS communication.

TODO: write SMS servlet that triggers this protocol as a parser/handler
TODO: map the current messages implementation to the new messages framework

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@69903 0d108963-8a47-0410-809b-ebd9fb3b3407