Xemex ReMI - Fix to LoadProfile readout.
The Xemex ReMI device sends its loadProfile without gaps - if no data is available (e.g. no data because of a power down), the device will fill its loadProfile with dummy data. This is done because some head-end systems can not manage loadProfiles with gaps. EIServer however can manage, so during parsing of the intervalData, the dummy data will be filtered out  (and thus not saved in EIServer, making it a proper 'missing' interval).

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@73492 0d108963-8a47-0410-809b-ebd9fb3b3407