EISERVERSG-3260 - Remove unused message editors and builders from MDW
Removed some unused message builders, 
Moved the other message builders to the protocols. These are still used to parse the message content in the protocol.
Also removed the interfaces (and their usages) that enabled these special messages in 8.11. This functionality has been completely replaced by the new message framework.
Also removed the editors that showed the special messages in the 8.11 UI.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@82724 0d108963-8a47-0410-809b-ebd9fb3b3407