COMMUNICATION-999 - Migrate EIWebPlus servlet from 8.X to 9.1
Added method findLikeProtocolProperty() to the DeviceFactory. This searches for devices with a given property value, allowing wildcards.
Added the Hawkeye modbus protocol to the capability and security mappings so it can be used by the adapters.

Added EIWebPlusMessageConverter to translate the new message specs into legacy XML (for the EIWebPlus servlet and RTU+Server)
Added a EIWebPlusDialect that contains the properties for the RtuServer protocol for EIWebPlus

Added VirtualLoadProfile property to the modbus protocols

Added protocol RtuServer that is a basically a placeholder for the RTU+Server concentrator device in EIServer.
It never executes anything, it's only there as a placeholder to provide the properties, security settings and messages.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@82733 0d108963-8a47-0410-809b-ebd9fb3b3407