EISERVERSG-3212 - Security Sets Slave devices - inherit from Master
Added an extra encryption and authentication level to the existing V2 slave protocols: InheritedAuthenticationDeviceAccessLevel and InheritedEncryptionDeviceAccessLevel.
This is a placeholder level that indicates that the slave protocol can inherit the security properties of the master device, instead of (re)specifying these properties them self.

I adjusted the (smart)meterprotocoladapter so it provides these same 2 levels for every legacy (V1 and V11) protocol that has slave capability.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@83235 0d108963-8a47-0410-809b-ebd9fb3b3407