COMMUNICATION-1011 - Create a confluence tool to generate message documentation
The client used for the SOAP services was imported using the provided WSDL file.

Use SyncFullMessageOverview to update the overview of all messages in confluence.
This looks for a page called "All messages" in space "PRTCL".

Use SyncMessagesForProtocol to update the overview of all messages supported by 1 protocol in it's release notes on confluence.
This looks for a page that has the exact protocol description name, in space "PRTCL".

Note that you have to provide credentials to do the update.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@83343 0d108963-8a47-0410-809b-ebd9fb3b3407