EISERVERSG-3243 - Rework of remote comserver according to 9.1 ComServer model
Completed the remote ComServer so that it can be fully executed without database access.

Extended the ComServer DAO to enable some functionality for the Remote ComServer that would otherwise require database access.
E.g.: get versioned properties (can not be cached)
Added the necessary tags (@XmlType and @XmlAttribute etc) to serialize all necessary fields used by the remote side. 
Message mappings and security mappings are now accessed using the DAO (for the remote comserver)
Renamed getters and setters for fields so that they have the proper format (e.g. getField returns field)

Replaced the transaction executor for the remote comserver by a dummy one, since we don't have database access. This one simply executes the doExecute() method.
Replaced the format provider with a default one.
Replaced the format preferences with a fixed one.
Replaced the translator with a remote compatible one (that uses the DAO to fetch translations from the DB)

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@83782 0d108963-8a47-0410-809b-ebd9fb3b3407