EISERVERSG-3541 - Limit Load-Profile read-out for master/slave
Changed the way combined LPs are read out: each part is read out individually, by its own comtask, defined on the proper device.
Note that since the protocols have no selective access based on channels yet, some data will be read multiple times in case of a combined LP.
I chose to not implement an optimization in the framework for this, because the different LP parts now belong to different comtasks, which can have different protocol task settings and are executed individually, with their own logging and scheduling.

- Changed OfflineDevice to return all LPs. 
The multiple parts of a "combined LP" (LP on master and LP on slave of the same LP type)  are now returned as multiple, individual LP's (instead of just 1 combined LP)
The effect of this change is that every LP on every device must be read out using an individual comtask.
A combined LP is no longer fully read by just 1 task on the master. Every device (master and slaves) must have their own LP comtask.

- When creating LoadProfileReaders and adding them to the LP ComCommands, only the relevant channels are included.
These are the channels of the actual device only. In case of a "combined" LP, the channels of the other devices are excluded.

- Added method removeUnwantedChannels() to the ComCommands that read out load profile data.
This filters out any channel intervals that were not requested.
Note that this method will no longer be necessary when the protocols implement selective access based on channels.

- Added a getter for the LP obiscode on all the LP identifiers. This can now be used to compare LP identifiers.

- Slightly changed the UtcLoadProfileDataWriter. It will only store intervals for the direct channels of the LP. 
It no longer stores interval data for the external, virtual channels (that are defined on slave LPs).

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@88973 0d108963-8a47-0410-809b-ebd9fb3b3407