COMMUNICATION-1182 - Add more info in Unknown Device Logging for EIWeb devices
Adjusted the inbound framework so that the protocol implementations can provide extra information, to further identify the inbound device.
This additional info is logged in the unknown devices logging window, if it turns out the device identifier cannot be resolved.

At the moment, only the EIWebBulk has a proper implementation: it provides the device ID, the serial number and the IP address as additional info.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@90052 0d108963-8a47-0410-809b-ebd9fb3b3407