EISERVERSG-3888 - Physical slave devices cannot choose dialect properties
Temporary workaround untill ticket EISERVERSG-3888 is fixed.
Physical slave devices now always use the NodeAddress property as value for the ServerUpperMacAddress, regardless of the value of the ServerUpperMacAddress property.
The NodeAddress is a general property, specified on the physical slave itself.
The ServerUpperMacAddress property is (currently) a dialect property, specified on the gateway device... This is useless end needs a fix.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@91383 0d108963-8a47-0410-809b-ebd9fb3b3407