COMMUNICATION-1246 - Change Set Encryption Keys message properties to be of password type
Changed the propertyspec of some message attributes that hold a key or password. It is now of type Password, so the value is stored encrypted in the DB.

Note that existing 9.1.x systems that already have such messages will now throw an application exception when showing them in the message overview.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@91766 0d108963-8a47-0410-809b-ebd9fb3b3407