COMMUNICATION-1303 - WPDU errors
Receiving an invalid TCP frame (invalid source or destination) is now better handled. Instead of stopping the session, the full frame is read and ignored. 
After this, we attempt to read out the next full frame, so the protocol session can continue.

This fix only applies for newer protocols that don't use polling in the connection layer.

git-svn-id: http://source.eict.vpdc/svn/labo/protocols/trunk@92955 0d108963-8a47-0410-809b-ebd9fb3b3407