COMMUNICATION-1514 - Protocol executor fails to read dummy meter with security suite 0, profile 5-3
Now only validating that the received FC is higher than the previously received FC.
Should not be strictly +1.
This was changed for the Beacon protocol, the IDIS P2 protocols and the AM540 DSMR protocol.

See specs:

DSMR5.0:
The frame counter used in the messages send by the meter should be incremented by
exactly one for each frame send.
The CS/gateway will validate the Frame Counter and check that the received Frame
Counter is higher than in a previous message.

IDIS P2:
A message received on a specific channel is discarded if the frame counter in the received
message is smaller or equal to the frame counter expected by the receiver on this channel.