COMMUNICATION-1474 - AM540 IDIS: Inefficient use of block size
Added optimization for the readout of the LP interval and the LP channel units: they are now  read out in bulk.
Same goes for the serial number and the clock.
The clock (time difference) is now cached in the protocol instance, so it is read out only once.  This is applied for all V2 protocols.