COMMUNICATION-1551 - FW Upgrade Fails when frame counter in received message is smaller or equal to frame counter in previously received message
Adjusted the G3RespondingFrameCounterHandler impl: in case of invalid framecounter, a different exception type is thrown.
It is recognized by the SecureConnection layer that will then disregard the APDU, and listen for the next APDU.
This is now applied for the AM540 V2 protocols (IDISP2 and DSMR).

Note that the default behaviour for other protocols (in case of invalid frame counter) is still to abort the session.