COMMUNICATION-1582 - GatewayTCPIPConnection should let the SocketErrors through
Removed class LowerLayerCommunicationException and its handling

Now added the logic in the ComServer framework instead of the protocols. The framework now checks which kind of ConnectionCommunicationException is thrown by the protocol.
Any timeout/ciphering/fatal protocol application error will result in skipping the tasks of the current physical slave device. The tasks of the next physical slave devices can still be executed, as long as the connection is intact.

Also fixed a bug in the ComServer where comtasks were marked OK while they should be marked 'not executed' because of a problem in an earlier comtask for the same device.