COMMUNICATION-1582 - GatewayTCPIPConnection should let the SocketErrors through
Made the difference between a timeout that means that your connection is also broken, and a timeout that still has an intact connection.
Protocols can indicate it in their implementation of CommunicationSessionProperties.
This is essential when reading out physical slave devices that have a common gateway (and thus connection task).

The default behavior for protocols is that a timeout also breaks the connection.