CXO-1334 - View all reading qualities on channel readings (and remove Interval flags column)
Replaced all usages of class ProfileStatus by Reading Qualities.
The ProfileStatus class is deleted.
Usage of the IntervalStateBits and eiStatus in IntervalData/IntervalValue has been depricated, use reading qualities instead

A mapping between the old interval flags and reading quality CIM codes is implemented in IntervalFlagMapper.
This is used when the old protocols return profile data that have interval flags.

Added translation for the reading quality CIM code - category.
Added some custom CIM codes for interval flags that could not be mapped

Reading qualities are now included in the REST layer instead of interval flags so they can be shown in the UI.

Note that the recordspecs of an intervalreading has not changed. The second column (profile status) is still present, it is fixed to 0L.