Beacon 3100 : also release if pre-established prop is overridden.
If the configuration states a pre-established public client association
in combination with an authenticated frame counter, the implementation
should override this since this combination is not possible. In this
case we'll use a non-pre-established association regardless of what the
configuration says to avoid weird errors. However, we also need to
release in that case.