COMMUNICATION-1644 - EVN HSM integration
Extended the AM540 IDIS protocol and the Beacon3100 protocol so the datatransport security (encryption, authentication of APDUs) and the HLS security is delegated to an HSM using the ProtocolService API. This is only the case when using 'irreversible keys'.
In the case of plain keys or keys that can be reversed by the HSM, the normal protocol functionality is still used.