CXO-2951 - Integrate existing DeviceProtocol classes in Connexo code base
Fixed bug for estimators that used a property of type TimeDuration. See also CXO-6088.
I re-introduced the propertyspec of type TimeDuration (additionally to the new Duration and TemporalAmount property spec types) so that this can be used in the estimators (and anywhere else in Connexo).
I also moved the new Duration and TemporalAmount property spec types to the properties module (instead of mdc) so that anyone in connexo can use these.